requirements.txt
numpy
pandas
kaggle
seaborn
librosa
colorama
langchain
streamlit
matplotlib
scikit-learn
python-dotenv
-e .